Divide tts handlers
