+++ src/handlers/tts/__init__.py
@@ -1,2 +1,14 @@
-from .tts import *
+from .tts import TTSHandler
+from .custom_handler import CustomTTSHandler
+from .espeak_handler import EspeakHandler
+from .gtts_handler import gTTSHandler
+from .elevenlabs_handler import ElevenLabs
+
+__all__ = [
+    "TTSHandler",
+    "CustomTTSHandler",
+    "EspeakHandler",
+    "gTTSHandler",
+    "ElevenLabs"
+]
