+++ src/handlers/tts/custom_handler.py
@@ -1,0 +1,37 @@
+from subprocess import check_output
+
+from .tts import TTSHandler
+from ...utility.system import get_spawn_command 
+
+
+class CustomTTSHandler(TTSHandler):
+    def __init__(self, settings, path):
+        self.settings = settings
+        self.path = path
+        self.key = "custom_command"
+        self.voices = tuple()
+
+    @staticmethod
+    def requires_sandbox_escape() -> bool:
+        """If the handler requires to run commands on the user host system"""
+        return True
+
+    def get_extra_settings(self) -> list:
+        return [{
+            "key": "command",
+            "title": _("Command to execute"),
+            "description": _("{0} will be replaced with the model fullpath"),
+            "type": "entry",
+            "default": ""
+        }]
+
+
+    def is_installed(self):
+        return True
+
+    def play_audio(self, message):
+        command = self.get_setting("command")
+        if command is not None:
+            self._play_lock.acquire()
+            check_output(get_spawn_command() + ["bash", "-c", command.replace("{0}", message)])
+            self._play_lock.release()
