+++ src/handlers/tts/tts.py
@@ -1,7 +1,6 @@
 from abc import abstractmethod
 from typing import Callable
 
-from gtts import gTTS, lang
 from subprocess import check_output
 import threading 
 import time
@@ -105,174 +104,3 @@ class TTSHandler(Handler):
         self.set_setting("voice", voice)
 
 
-class gTTSHandler(TTSHandler):
-    key = "gtts"
-   
-    def get_voices(self):
-        if len(self.voices) > 0:
-            return self.voices
-        x = lang.tts_langs()
-        res = tuple()
-        for l in x:
-            t = (x[l], l)
-            res += (t,)
-        self.voices = res
-        return res
-
-    def save_audio(self, message, file):
-        voice = self.get_current_voice()
-        if not self.voice_available(voice):
-            voice = self.get_voices()[0][1]
-        tts = gTTS(message, lang=voice)
-        tts.save(file)
-
-
-class EspeakHandler(TTSHandler):
-    
-    key = "espeak"
-
-    @staticmethod
-    def requires_sandbox_escape() -> bool:
-        """If the handler requires to run commands on the user host system"""
-        return True
-
-    def get_voices(self):
-        if len(self.voices) > 0:
-            return self.voices
-        if not self.is_installed():
-            return self.voices
-        output = check_output(get_spawn_command() + ["espeak", "--voices"]).decode("utf-8")
-        # Extract the voice names from the output
-        lines = output.strip().split("\n")[1:]
-        voices = tuple()
-        for line in lines:
-            spl = line.split()
-            voices += ((spl[3], spl[4]),)
-        self.voices = voices
-        return voices
-
-    def play_audio(self, message):
-        self._play_lock.acquire()
-        check_output(get_spawn_command() + ["espeak", "-v" + str(self.get_current_voice()), message])
-        self._play_lock.release()
-
-    def save_audio(self, message, file):
-        r = check_output(get_spawn_command() + ["espeak", "-f", "-v" + str(self.get_current_voice()), message, "--stdout"])
-        f = open(file, "wb")
-        f.write(r)
-
-    def is_installed(self):
-        if not can_escape_sandbox():
-            return False
-        output = check_output(get_spawn_command() + ["whereis", "espeak"]).decode("utf-8")
-        paths = []
-        if ":" in output:
-            paths = output.split(":")[1].split()
-        if len(paths) > 0:
-            return True
-        return False
-
-class CustomTTSHandler(TTSHandler):
-    def __init__(self, settings, path):
-        self.settings = settings
-        self.path = path
-        self.key = "custom_command"
-        self.voices = tuple()
-
-    @staticmethod
-    def requires_sandbox_escape() -> bool:
-        """If the handler requires to run commands on the user host system"""
-        return True
-
-    def get_extra_settings(self) -> list:
-        return [{
-            "key": "command",
-            "title": _("Command to execute"),
-            "description": _("{0} will be replaced with the model fullpath"),
-            "type": "entry",
-            "default": ""
-        }]
-
-
-    def is_installed(self):
-        return True
-
-    def play_audio(self, message):
-        command = self.get_setting("command")
-        if command is not None:
-            self._play_lock.acquire()
-            check_output(get_spawn_command() + ["bash", "-c", command.replace("{0}", message)])
-            self._play_lock.release()
-
-class ElevenLabs(TTSHandler):
-    key = "elevenlabs"
-    def get_extra_settings(self) -> list:
-        return [
-            {
-                "key": "api",
-                "title": _("API Key"),
-                "description": _("API Key for ElevenLabs"),
-                "type": "entry",
-                "default": ""
-            },
-            {
-                "key": "voice",
-                "title": _("Voice"),
-                "description": _("Voice ID to use"),
-                "type": "entry",
-                "default": "21m00Tcm4TlvDq8ikWAM"
-            },
-            {
-                "key": "model",
-                "title": _("Model"),
-                "description": _("Name of the model to use"),
-                "type": "combo",
-                "values": (("eleven_turbo_v2_5", "eleven_turbo_v2_5"), ("eleven_multilingual_v2", "eleven_multilingual_v2")),
-                "default": "eleven_turbo_v2_5"
-            },
-            {
-                "key": "stability",
-                "title": _("Stability"),
-                "description": _("stability of the voice"),
-                "type": "range",
-                "min": 0,
-                "max": 1,
-                "round-digits": 2,
-                "default": 0.50
-            },
-            {
-                "key": "similarity",
-                "title": _("Similarity boost"),
-                "description": _("Boosts overall voice clarity and speaker similarity"),
-                "type": "range",
-                "min": 0,
-                "max": 1,
-                "round-digits": 2,
-                "default": 0.75
-            },
-            {
-                "key": "style_exaggeration",
-                "title": _("Style exaggeration"),
-                "description": _("High values are reccomended if the style of the speech must be exaggerated"),
-                "type": "range",
-                "min": 0,
-                "max": 1,
-                "round-digits": 2,
-                "default": 0
-            },
-
-        ]
-
-    @staticmethod
-    def get_extra_requirements() -> list[str]:
-        return ["elevenlabs"]
-
-    def save_audio(self, message, file):
-        from elevenlabs.client import ElevenLabs 
-        from elevenlabs import save
-        from elevenlabs.types import VoiceSettings
-        client = ElevenLabs(api_key=self.get_setting("api"))
-        audio = client.generate(text=message, voice=self.get_setting("voice"), model=self.get_setting("model"),
-                                voice_settings=VoiceSettings(stability=self.get_setting("stability"), similarity_boost=self.get_setting("similarity"), style=self.get_setting("style_exaggeration")))
-        save(audio, file) 
-        return
